+++ src/main/java/com/takeshi/library/service/UserService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface UserService {
     List<UserEntity> searchUsers(String keyword);       // 一覧・検索
-    UserEntity findById(Long id);                       // IDで取得（編集・削除用）
+    UserEntity findById(Integer id);                       // IDで取得（編集・削除用）
     void insert(UserEntity userEntity);                       // 新規登録
     void update(UserEntity userEntity);                       // 更新
-    void softDelete(Long id);                     // 論理削除
+    void softDelete(Integer id);                     // 論理削除
     UserEntity findByEmail(String email);
 }
